Can handle blocks that create multiple nodes
This was tricky... The API for the builder expects you to always use the
method_missing approach and if you stray from there, you have to work in
the traditional Nokogiri API. That's fine, but it wasn't immediately
clear that the `parent` attribute on the builder is intended to be
mutable. I had to do a fair amount of re-reading of the implementation
of the Builder code to figure that out. I wish that the builder exposed
an API for something like `nest_under(tagname)` that would swap the
parent and allow you be a little more stateful.