Convert block-entityKeys to strings
This is ... fiddly. The `convertToRaw` function that is built into
DraftJS generates numeric entity keys in the `entityRanges`, but the
resultant value in the `entityMap` is a string. To match with that
expectation and prevent surprises, we will use this behavior ... but
it may need to evolve to a configuration