Handle entities/styles on the same node correctly
We noticed that some entities/styles did not play nicely together.
This changes the renderer function APIs to receive "content"/Node
objects that can wrap themselves in another node. It may be necessary
at some point to provide another API for nesting the other way, but
at the moment, the API is flexible enough that you can do it without
one.