Enable parsing FromHtml -> Draftjs
What a journey! This commit introduces the ability to parse from HTML
into DraftJS and returns the raw Draftjs content.

It's ... rather opinionated. This isn't necessarily intentional, but
rather mostly due to simplifying assumptions that made it easier to get
to a v0.1.0 for this. Definitely consider this "beta" software -
feedback welcome.

Some known limitations:

1. The node lists and their handling are pretty rigid.

   This could be a limitation for you if, say, you wanted to handle
   `table` elements in a different way (right now they're _kinda_ treated
   as list-parents).

1. Certain "invalid HTML" cases may not do what you want

   - One such example is if you have an "inline style" tag (e.g. `b`)
     wrapping a block node - for example. The below is technically
     invalid, and the wrapping `b` tag will not bold the content inside the
     `p` tag.

     ```
     <div>
       <b>
         <p>I will not be bold</p>
       </b>
     </div>
     ```