[FromHtml] Track entities on each character
This borrows from the internal DraftJS implementation the idea of
keeping track of each character with its associated entities, rather
than trying to manipulate indexes. Then, when we need/want to convert
the in-memory representation, we do the index calculations _then_, and
not before. This prevents a whole class of errors (off-by-1s, etc)
because it let's us apply those entities _once_ and then the character
is forever set with that entity.

This could(should?) likely go further to include the inline styles on
each character, but getting this idea in place for the entities first
solves a consumer issue.