More faithfully handle intent when parsing tables/divs
This introduces the concept of a "FLUSH_BOUNDARY" internally. These are
tags that, when encountered inside another tag, denote the start of a
new block. That is, while parsing, if you encounter a `tr` inside a
`tr`, you probably want to flush whatever's pending before proceeding.
The thought here, is to more faithfully handle "weird" HTML based on the
user's intent. _Most_ of our users are not writing their own HTML - it's
generated for them by a tool - so `table` tags, `div` tags, etc are
more likely to be artifacts of the tool, and not intentional markup.

Prior to this change, nested tables and nested divs had to be handled in
a much more "ad-hoc" fashion. This change allows us to unify that, at
the risk of being more opinionated. That is, the library, with this
change, is taking a stance that the default behavior to try and devine
user intent with the markup, _not_ be 100% faithful to the HTML. There
may come a day when that changes or is configurable, but today is not
that day.