[FromHTML] Introduce `squeeze_whitespace_blocks` option
This option allows you to specify that you do _not_ want to keep
"insignficant" whitespace from the HTML and tries to go along with
"intuiting" what the user intended. In the event you have many `br` tags
or `div` tags in a row with nothing in them, this option _ignores_ the
multitude of tags and "simply" uses the DraftJS block-boundary to
establish a deliniation between the "before" content" and the "after"
content.