Bump version to 0.28.0 for arbitrary depths in ToHtml
This creates 'multi-push' logic for nested lists to handle cases where
draftjs has non-incremental depth jumps. We find this may happen with
copy-pasted content. In this algorithm, we found that if you have a >1
depth jump, li elements must be created for the intermediate depth
levels as well as list parents. To maintain the expectations the calling
code has (block is a li iteself), we cannot create the final depth list
item. Which is why we finish the method with a dangling push parent.
@LBPeterson