Handle newly discovered infinite loop issue
This fixes a newly discovered issue (documented in a test) that was
causing the to_html conversion to run in an infinite loop. Solving this
surgically was not working with the old algorithm, so this commit
"simplifies" the depth handling code by being as dumbly-procedural as
possible. Rather than trying to get cute with naming and/or bad
abstraction, this straight-up uses a series of `if`s to get the job
done.

There's one "breaking" change, related to sibling non-list nodes:
they are now included _in the sibling list element at the same depth_.
This was probably the intent all along, but it was difficult to discern
with the previous algorithm.

I'm still incredibly open to other ideas on how to overhaul this. I feel
like there _should_ be some kind of elegant way to map here, but I can't
seem to get it right.