Add concept of NullContent & Draftjs.safe_parse
We found that in our application we often wanted a `NullObject` style actor
rather than `nil` to avoid the implicit type-checks of `.nil?` and
special-casing behavior. We did not want to break the current behavior of
`.parse` so opted to introduce `.safe_parse`, which captures any exception and
guarantees a benign return value. This also marks `Draftjs.parse` &
`Draftjs.safe_parse` as part of the public API of this gem, as noted in the
README.

Previously, `DraftjsHtml::Draftjs.parse` raised and the specific exception was
unknown, so you essentially had to rescue everything. Consumers can now catch
`DraftjsHtml::InvalidRawDraftjs` for errors specific to the `raw` input.

For the future, we may still consider pulling out the `Draftjs` submodule,
leaving this gem’s responsibility solely focused on to-and-from HTML.